requirements.txt
PyPDF2==3.0.1
pdfplumber==0.11.0
reportlab==4.0.9
Pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
